reorganize code, removing unified directory
