+++ src/stub.d.ts
