+++ src/stub.js
